Restructured original project, and appended new React UI
Appended the new user-interface to the forked repo of Yoichiro Tanaka.
To do so, the forked repo structure is changed: the app folder is
renamed to 'src'. The 'scripts' folder is renamed to 'js' and sub-
divided in 'model', 'controller', 'view' and 'utils'. Scripts who
empower communcation with the Native Client are stored in 'model'.
All react components (for the user-interface) are placed in 'view'.
The 'background' script can be found in 'controller'. The 'window.js'
script is renamed to 'app.js' and placed within 'controller'. Unused
scripts, such as 'index.js', 'chromereload.js' and 'index.js' are
placed within 'utils'.