Auto close Drawer on blur, fixed form visibility when Drawer is open
When the Drawer pane is open, the Drawer will be automatically closed
as soon the user clicks anywhere in the form-div.

In case the window is very small, the Drawer pane will overlay the
form field (when opened). In case the window is large enough, the
Drawer pane will shift the form field.