Fix kex methods (#1)
* Update sftp_thread.cc

Set our preferred keys to the default OpenSSH list. Because libssh2 is somewhat outdated in this category, it will only support `diffie-hellman-group-exchange-sha256` for the moment, but this _should_ help fix errors like #131. Still some more changes to make before this can be PR'd, but should be a lot closer. I hope.

* Update English messages

Dutch is next!

* Bijgewerkte Nederlandse berichten

Hoewel ik Google Translate heb gebruikt...

(For you English folk, that's an update to the Dutch messages, as promised in the previous commit.)

* Update sftp_thread.h

Gotta declare the method before we can define it.